parameters check

+++ server/backup-db.sh
@@ -1,13 +1,27 @@
-user=github
-password=$REPO_PASSWORD
+#!/bin/bash
+
 db_container=pandora_db_1
+user=github
+
+vars_expected="$vars_expected REPO_PASSWORD";        vars_found="$vars_found${REPO_PASSWORD+ REPO_PASSWORD}"
+vars_expected="$vars_expected STORAGE_BASE_URI";     vars_found="$vars_found${STORAGE_BASE_URI+ STORAGE_BASE_URI}"
+vars_expected="$vars_expected STORAGE_BACKUPS_REPO"; vars_found="$vars_found${STORAGE_BACKUPS_REPO+ STORAGE_BACKUPS_REPO}"
+
+if [ "$vars_found" != "$vars_expected" ]
+then
+    echo "Required environment varibles:$vars_expected" >&2
+    echo "Found:$vars_found" >&2
+    exit 1
+fi
+password=$REPO_PASSWORD
 archive_name="$db_container-$(date +'%y%m%d_%H%M%S').tgz"
 local_path=/tmp/$archive_name
-remote_path="$REPO_URI/$archive_name"
+remote_path="$STORAGE_BASE_URI/$STORAGE_BACKUPS_REPO/$archive_name"
+storage_api_uri="$STORAGE_BASE_URI/api/storage/$STORAGE_BACKUPS_REPO"
 
 docker exec $db_container sh -c 'rm -r /tmp/bkp* ; mongodump --username=$MONGO_INITDB_ROOT_USERNAME --password=$MONGO_INITDB_ROOT_PASSWORD --out=/tmp/bkp && tar czvf /tmp/bkp.tgz /tmp/bkp && rm -r /tmp/bkp'
 docker cp $db_container:/tmp/bkp.tgz $local_path
-curl -u github:$password -T $local_path "$remote_path"
+curl -u $user:$password -T $local_path "$remote_path"
 
 # https://www.jfrog.com/confluence/display/JFROG/Artifactory+REST+API#ArtifactoryRESTAPI-FolderInfo
